webapp/requirements.txt
confluent-kafka
protobuf
psycopg2-binary
requests
Faker
Flask
